feat: implement Permask class for flexible permission management
