+++ src/index.ts
@@ -2,3 +2,4 @@ export * from "./permask";
 export * from "./utils/bitmask";
 export * from "./constants/permission";
 export * from "./integrations/express";
+export * from "./permask-class";
